backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend"
requires-python = "==3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "python-multipart",
    "transformers",
    "torch",
    "ffmpeg-python",
    "numpy",
    "pytest",
    "pytest-asyncio",
    "langchain-core>=0.3.67",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "langchain-ollama>=0.3.3",
    "mlx-whisper; sys_platform == 'darwin'",
]

[tool.setuptools.packages.find]
exclude = ["templates*", "temp_audio*", "__pycache__*"]
